# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend Go untuk manga library: scan/move folder, bookmark, translate, bug report, duplikat.\nLogin lewat POST /login, lalu klik Authorize dan isi `Bearer \u003caccess_token\u003e`.",
        "title": "Manga Library API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/upload/folder": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "Upload folder manga ke SRC_DIR",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nama folder",
                        "name": "folder_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "multi",
                        "description": "Halaman-halaman (boleh banyak)",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Model.BaseResponseModel"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "type": "object",
                                            "properties": {
                                                "files_skipped": {
                                                    "type": "integer"
                                                },
                                                "files_written": {
                                                    "type": "integer"
                                                },
                                                "folder_name": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Model.BaseResponseModel": {
            "type": "object",
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "Data": {},
                "HeaderMessage": {
                    "type": "string"
                },
                "Message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: `Bearer \u003caccess_token\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}